+++ velox/experimental/stateful/rank/RowTimeDeduplicateRanker.h
@@ -57,12 +57,12 @@ class RowTimeDeduplicateRanker : public exec::Operator,
 
   void open(StreamOperatorStateHandler* stateHandler) override;
 
-  RowVectorPtr processElements(uint32_t key, RowVectorPtr input) override;
+  RowVectorPtr processElements(int64_t key, RowVectorPtr input) override;
 
   void close() override;
 
  private:
-  std::shared_ptr<ValueState<uint32_t, int64_t, RowVectorPtr>> state_;
+  std::shared_ptr<ValueState<int64_t, int64_t, RowVectorPtr>> state_;
   int64_t minRetentionTime_;
   int rowtimeIndex_;
   bool generateUpdateBefore_;
