+++ velox/experimental/stateful/state/HeapKeyedStateBackend.h
@@ -30,7 +30,7 @@ class HeapKeyedStateBackend : public KeyedStateBackend {
   std::shared_ptr<ListState<uint32_t, int64_t, RowVectorPtr>>
   getOrCreateListState(StateDescriptor& stateDescriptor) override;
 
-  std::shared_ptr<ValueState<uint32_t, int64_t, RowVectorPtr>>
+  std::shared_ptr<ValueState<int64_t, int64_t, RowVectorPtr>>
   getOrCreateValueState(StateDescriptor& stateDescriptor) override;
 
   std::shared_ptr<InternalTimerService<int64_t, int64_t>> createTimerService(
@@ -45,9 +45,9 @@ class HeapKeyedStateBackend : public KeyedStateBackend {
   std::shared_ptr<MapState<uint32_t, int, uint32_t, RowVectorPtr>>
   getOrCreateRankMapState(StateDescriptor& stateDescriptor) override;
 
-  virtual std::shared_ptr<InternalTimerService<uint32_t, TimeWindow>>
+  virtual std::shared_ptr<InternalTimerService<int64_t, TimeWindow>>
   createGroupWindowAggTimerService(
-      Triggerable<uint32_t, TimeWindow>* triggerable) override;
+      Triggerable<int64_t, TimeWindow>* triggerable) override;
 
   void snapshot(
       int64_t checkpointId,
